+++ src/index.ts
@@ -1,2 +1,3 @@
 import "./shaders/landscapeShader.js";
 export * from "./renderer.js";
+export * from "./worldSpaceCompositor.js";